c-brahms/requirements.txt
music21==2.1.2
nose==1.3.7
nose-parameterized==0.6.dev0
numpy==1.11.1
pandas==0.16.2
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
vis-framework==2.4.1
